Fixing Game 1: Frog Feast
I decided to make the initial 'hungryhungryfrog' game as the first option in my new game that involves the other mini games. I decided to also fix some of the issues i had with it like the winning screen and losing screen. I also added an exit button at the bottom right to give more options throughout the game experience. There's also an imrpovement in the visuals and some functions of this first game.