Game 2: Frog Pong
This game involves a user-controlled paddle (right paddle) and a AI controlled paddle (left paddle). I made the AI controlled paddle have moments where it loses the ball so the user could have opportunities to score. It's not perfect as I think the AI-controlled paddle still seems pretty good. The left paddle seems to also jitter a lot, im guessing it has to do with the ai-controlled left paddle that it hits the up and down arrow a lot which makes it look jittery. I haven't found a way to fix it yet but I would like to smoothen the movements in the future.