Game 3: Frog Shooter
This game was inspired by those space shooter arcade games I used to play back then, i wanted to create different tiers of games, and in this one I added obstacles that made the game a bit more difficult. First having to shoot falling 'froggies' 5x more to kill them and gain points. I also added the obstacle that when the froggies hit the bottom, the user loses points and has to restart the game. This adds a bit more difficulty in the game and adds something different from the previous 2 games.