Game 4: Frog Leap
For this game, I was told by my gf to do a frog jumping on lily pads. I thought about calling it 'Frog Leap'. The functions was a bit hard to come by but I managed to do something funtional. Inspired by the Mario Bros with the platforming game, as well as Flappy Bird with the game mechanics of jumping and arrow keys, it was fun having this work the way it is functioning right now. Initally the game was really hard to win because it immediately played after pressing the game in the menu screen, so I thought about putting a 'intro' screen for this game where it tells the user to press the space bar so it gives the player some room to breathe before starting the game.